requirements.txt
annotated-types==0.6.0
anyio==3.7.1
click==8.1.7
dnspython==2.4.2
fastapi==0.104.0
h11==0.14.0
httptools==0.6.1
idna==3.4
pydantic==2.4.2
pydantic_core==2.10.1
pymongo==4.5.0
python-dotenv==1.0.0
PyYAML==6.0.1
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.23.2
uvloop==0.18.0
watchfiles==0.21.0
websockets==11.0.3
